+++ lib/rubomop.rb
@@ -2,6 +2,7 @@
 
 require "active_support/core_ext/array"
 require "active_support/core_ext/string/inflections"
+require "active_support/core_ext/string/inquiry"
 require "awesome_print"
 require "fileutils"
 require "literal"